Add files via upload
A straightforward online application was created utilising JavaScript, CSS, and HTML to create this password generator. Its goal is to produce secure, random passwords that may be applied to several internet accounts and services.
The Password Generator creates passwords with a predetermined length, a mix of upper- and lowercase letters, digits, and special characters using a straightforward algorithm. Depending on their preferences, users can alter the length and difficulty of the passwords.
Users may easily generate passwords quickly and effectively thanks to the Password Generator's easy-to-use UI.
Please be aware that although this password generator is designed to generate secure passwords, it is not advised to use it for sensitive or important accounts.